Add a variety of order-based methods: first, last, get_gt, get_gte, get_lt, get_lte
