+++ src/lib.rs
@@ -443,6 +443,35 @@ impl Minimum for &str {
 /// more quickly, but the efficiency will be lower. Values that are
 /// extremely high may cause undesirable memory usage because it will
 /// take more time to fill up each thread-local garbage segment.
+///
+/// # Examples
+///
+/// ```
+/// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+///
+/// // insert and remove atomically returns the last value, if it was set,
+/// // similarly to a BTreeMap
+/// assert_eq!(tree.insert(1, 10), None);
+/// assert_eq!(tree.insert(1, 11), Some(10));
+/// assert_eq!(tree.remove(&1), Some(11));
+///
+/// // get also functions similarly to BTreeMap, except it
+/// // returns a cloned version of the value rather than a
+/// // reference to it, so that no locks need to be maintained.
+/// // For this reason, it can be a good idea to use types that
+/// // are cheap to clone for values, which can be easily handled
+/// // with `Arc` etc...
+/// assert_eq!(tree.insert(1, 12), None);
+/// assert_eq!(tree.get(&1), Some(12));
+///
+/// // compare and swap from value 12 to value 20
+/// tree.cas(1, Some(&12_usize), Some(20)).unwrap();
+///
+/// assert_eq!(tree.get(&1).unwrap(), 20);
+///
+/// // there are a lot of methods that are not covered
+/// // here - check out the docs!
+/// ```
 #[derive(Clone)]
 pub struct ConcurrentMap<K, V, const FANOUT: usize = 64, const LOCAL_GC_BUFFER_SIZE: usize = 128>
 where
@@ -610,6 +639,21 @@ where
     V: 'static + Clone + Send + Sync,
 {
     /// Atomically get a value out of the tree that is associated with this key.
+    ///
+    /// # Examples
+    /// ```
+    /// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+    ///
+    /// tree.insert(1, 1);
+    ///
+    /// let actual = tree.get(&0);
+    /// let expected = None;
+    /// assert_eq!(expected, actual);
+    ///
+    /// let actual = tree.get(&1);
+    /// let expected = Some(1);
+    /// assert_eq!(expected, actual);
+    /// ```
     pub fn get<Q>(&self, key: &Q) -> Option<V>
     where
         K: Borrow<Q>,
@@ -622,7 +666,178 @@ where
         leaf.get(key)
     }
 
+    /// Atomically get a key and value out of the tree that is associated with the key that
+    /// is lexicographically less than the provided key.
+    ///
+    /// This will always return `None` if the key passed to `get_lt` == `K::MIN`.
+    ///
+    /// # Examples
+    /// ```
+    /// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+    ///
+    /// tree.insert(1, 1);
+    ///
+    /// let actual = tree.get_lt(&0);
+    /// let expected = None;
+    /// assert_eq!(expected, actual);
+    ///
+    /// let actual = tree.get_lt(&1);
+    /// let expected = None;
+    /// assert_eq!(expected, actual);
+    ///
+    /// let actual = tree.get_lt(&2);
+    /// let expected = Some((1, 1));
+    /// assert_eq!(expected, actual);
+    /// ```
+    pub fn get_lt<Q>(&self, key: &Q) -> Option<(K, V)>
+    where
+        Q: Borrow<K> + ?Sized,
+    {
+        if key.borrow() == &K::MIN {
+            return None;
+        }
+
+        self.range(..key.borrow()).next_back()
+    }
+
+    /// Atomically get a key and value out of the tree that is associated with the key that
+    /// is lexicographically less than or equal to the provided key.
+    ///
+    /// # Examples
+    /// ```
+    /// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+    ///
+    /// tree.insert(1, 1);
+    ///
+    /// let actual = tree.get_lte(&0);
+    /// let expected = None;
+    /// assert_eq!(expected, actual);
+    ///
+    /// let actual = tree.get_lte(&1);
+    /// let expected = Some((1, 1));
+    /// assert_eq!(expected, actual);
+    ///
+    /// let actual = tree.get_lte(&2);
+    /// let expected = Some((1, 1));
+    /// assert_eq!(expected, actual);
+    /// ```
+    pub fn get_lte<Q>(&self, key: &Q) -> Option<(K, V)>
+    where
+        Q: Borrow<K> + ?Sized,
+    {
+        self.range(..=key.borrow()).next_back()
+    }
+
+    /// Atomically get a key and value out of the tree that is associated with the key
+    /// that is lexicographically greater than the provided key.
+    ///
+    /// # Examples
+    /// ```
+    /// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+    ///
+    /// tree.insert(1, 1);
+    ///
+    /// let actual = tree.get_gt(&0);
+    /// let expected = Some((1, 1));
+    /// assert_eq!(expected, actual);
+    ///
+    /// let actual = tree.get_gt(&1);
+    /// let expected = None;
+    /// assert_eq!(expected, actual);
+    ///
+    /// let actual = tree.get_gt(&2);
+    /// let expected = None;
+    /// assert_eq!(expected, actual);
+    /// ```
+    pub fn get_gt<Q>(&self, key: &Q) -> Option<(K, V)>
+    where
+        Q: Borrow<K> + ?Sized,
+    {
+        self.range((
+            std::ops::Bound::Excluded(key.borrow()),
+            std::ops::Bound::Unbounded,
+        ))
+        .next()
+    }
+
+    /// Atomically get a key and value out of the tree that is associated with the key
+    /// that is lexicographically greater than or equal to the provided key.
+    ///
+    /// # Examples
+    /// ```
+    /// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+    ///
+    /// tree.insert(1, 1);
+    ///
+    /// let actual = tree.get_gte(&0);
+    /// let expected = Some((1, 1));
+    /// assert_eq!(expected, actual);
+    ///
+    /// let actual = tree.get_gte(&1);
+    /// let expected = Some((1, 1));
+    /// assert_eq!(expected, actual);
+    ///
+    /// let actual = tree.get_gte(&2);
+    /// let expected = None;
+    /// assert_eq!(expected, actual);
+    /// ```
+    pub fn get_gte<Q>(&self, key: &Q) -> Option<(K, V)>
+    where
+        Q: Borrow<K> + ?Sized,
+    {
+        self.range((
+            std::ops::Bound::Included(key.borrow()),
+            std::ops::Bound::Unbounded,
+        ))
+        .next()
+    }
+
+    /// Get the minimum item stored in this structure.
+    ///
+    /// # Examples
+    /// ```
+    /// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+    ///
+    /// tree.insert(1, 1);
+    /// tree.insert(2, 2);
+    /// tree.insert(3, 3);
+    ///
+    /// let actual = tree.first();
+    /// let expected = Some((1, 1));
+    /// assert_eq!(actual, expected);
+    /// ```
+    pub fn first(&self) -> Option<(K, V)> {
+        self.iter().next()
+    }
+
+    /// Get the maximum item stored in this structure.
+    ///
+    /// # Examples
+    /// ```
+    /// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+    ///
+    /// tree.insert(1, 1);
+    /// tree.insert(2, 2);
+    /// tree.insert(3, 3);
+    ///
+    /// let actual = tree.last();
+    /// let expected = Some((3, 3));
+    /// assert_eq!(actual, expected);
+    /// ```
+    pub fn last(&self) -> Option<(K, V)> {
+        self.iter().next_back()
+    }
+
     /// Atomically insert a key-value pair into the tree, returning the previous value associated with this key if one existed.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+    ///
+    /// assert_eq!(tree.insert(1, 1), None);
+    /// assert_eq!(tree.insert(1, 1), Some(1));
+    /// ```
     pub fn insert(&self, key: K, value: V) -> Option<V> {
         loop {
             let mut guard = self.ebr.pin();
@@ -660,6 +875,16 @@ where
     }
 
     /// Atomically remove the value associated with this key from the tree, returning the previous value if one existed.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// let tree = concurrent_map::ConcurrentMap::<usize, usize>::default();
+    ///
+    /// assert_eq!(tree.remove(&1), None);
+    /// assert_eq!(tree.insert(1, 1), None);
+    /// assert_eq!(tree.remove(&1), Some(1));
+    /// ```
     pub fn remove<Q>(&self, key: &Q) -> Option<V>
     where
         K: Borrow<Q>,
@@ -819,6 +1044,10 @@ where
     ///
     /// But, you can be certain that any key that existed prior to the creation of this
     /// iterator, and was not changed during iteration, will be observed as expected.
+    ///
+    /// # Panics
+    ///
+    /// This will panic if the provided range's end_bound() == Bound::Excluded(K::MIN).
     pub fn range<R: std::ops::RangeBounds<K>>(
         &self,
         range: R,
@@ -834,7 +1063,10 @@ where
         let end = match range.end_bound() {
             std::ops::Bound::Unbounded => LeafSearch::Max,
             std::ops::Bound::Included(k) => LeafSearch::Eq(k),
-            std::ops::Bound::Excluded(k) => LeafSearch::Lt(k),
+            std::ops::Bound::Excluded(k) => {
+                assert!(k != &K::MIN);
+                LeafSearch::Lt(k)
+            }
         };
 
         let current = self.inner.leaf_for_key(LeafSearch::Eq(start), &mut guard);
@@ -933,7 +1165,7 @@ where
 impl<'a, K, V, const FANOUT: usize, const LOCAL_GC_BUFFER_SIZE: usize, R> DoubleEndedIterator
     for Iter<'a, K, V, FANOUT, LOCAL_GC_BUFFER_SIZE, R>
 where
-    K: 'static + Clone + Minimum + Ord + Send + Sync + fmt::Debug,
+    K: 'static + Clone + Minimum + Ord + Send + Sync,
     V: 'static + Clone + Send + Sync,
     R: std::ops::RangeBounds<K>,
 {
@@ -942,16 +1174,13 @@ where
             if self.next_index_from_back >= self.current_back.leaf().len() {
                 if !self.range.contains(&self.current_back.lo) || self.current_back.lo == K::MIN {
                     // finished
-                    // println!("None return 1");
                     return None;
                 }
-
-                // println!( "fetching new current_back with serach of Lt({:?})", self.current_back.lo);
 
                 let next_current_back = self
                     .inner
                     .leaf_for_key(LeafSearch::Lt(&self.current_back.lo), &mut self.guard);
-                assert_ne!(next_current_back.lo, self.current_back.lo);
+                assert!(next_current_back.lo != self.current_back.lo);
                 self.current_back = next_current_back;
 
                 self.next_index_from_back = 0;
@@ -962,7 +1191,6 @@ where
             }
 
             let offset_to_return = self.current_back.leaf().len() - (1 + self.next_index_from_back);
-            // println!( "offset: {} len: {} next_index_from_back: {}", offset_to_return, self.current_back.leaf().len(), self.next_index_from_back);
             let (k, v) = self
                 .current_back
                 .leaf()
@@ -970,19 +1198,15 @@ where
                 .unwrap();
 
             self.next_index_from_back += 1;
-            // println!("next_index_from_back is now {}", self.next_index_from_back);
             if !self.range.contains(k) {
-                // println!("None return 2 for k {:?}", k);
                 continue;
             } else {
-                // println!("returning Some({:?})", k);
                 return Some((k.clone(), v.clone()));
             }
         }
     }
 }
 
-// TODO: enforce Lt is never constructed with K::MIN
 enum LeafSearch<K> {
     // For finding a leaf that would contain this key, if present.
     Eq(K),
@@ -1808,7 +2032,7 @@ fn timing_tree() {
     let insert_elapsed = insert.elapsed();
     println!(
         "{} inserts/s, total {:?}",
-        (n * 1000) / u64::try_from(insert_elapsed.as_millis().max(1)).unwrap_or(u64::MAX),
+        (n * 1_000_000) / u64::try_from(insert_elapsed.as_micros().max(1)).unwrap_or(u64::MAX),
         insert_elapsed
     );
 
@@ -1818,8 +2042,18 @@ fn timing_tree() {
     let scan_elapsed = scan.elapsed();
     println!(
         "{} scanned items/s, total {:?}",
-        (n * 1000) / u64::try_from(scan_elapsed.as_millis().max(1)).unwrap_or(u64::MAX),
+        (n * 1_000_000) / u64::try_from(scan_elapsed.as_micros().max(1)).unwrap_or(u64::MAX),
         scan_elapsed
+    );
+
+    let scan_rev = Instant::now();
+    let count = tree.range(..).rev().count();
+    assert_eq!(count as u64, n);
+    let scan_rev_elapsed = scan_rev.elapsed();
+    println!(
+        "{} reverse-scanned items/s, total {:?}",
+        (n * 1_000_000) / u64::try_from(scan_rev_elapsed.as_micros().max(1)).unwrap_or(u64::MAX),
+        scan_rev_elapsed
     );
 
     let gets = Instant::now();
@@ -1829,194 +2063,7 @@ fn timing_tree() {
     let gets_elapsed = gets.elapsed();
     println!(
         "{} gets/s, total {:?}",
-        (n * 1000) / u64::try_from(gets_elapsed.as_millis().max(1)).unwrap_or(u64::MAX),
+        (n * 1_000_000) / u64::try_from(gets_elapsed.as_micros().max(1)).unwrap_or(u64::MAX),
         gets_elapsed
     );
-}
-
-#[test]
-fn concurrent_tree() {
-    let n: u16 = 1024;
-    let concurrency = std::thread::available_parallelism()
-        .map(std::num::NonZeroUsize::get)
-        .unwrap_or(8)
-        * 2;
-
-    let run = |tree: ConcurrentMap<u16, u16, 8>, barrier: &std::sync::Barrier, low_bits| {
-        let shift = concurrency.next_power_of_two().trailing_zeros();
-        let unique_key = |key| (key << shift) | low_bits;
-
-        barrier.wait();
-        for key in 0..n {
-            let i = unique_key(key);
-            assert_eq!(tree.get(&i), None);
-            tree.insert(i, i);
-            assert_eq!(tree.get(&i), Some(i), "failed to get key {i}");
-        }
-        for key in 0_u16..n {
-            let i = unique_key(key);
-            assert_eq!(tree.get(&i), Some(i), "failed to get key {i}");
-        }
-        for key in 0_u16..n {
-            let i = unique_key(key);
-            assert_eq!(
-                tree.cas(i, Some(&i), Some(unique_key(key * 2))),
-                Ok(Some(i)),
-                "failed to get key {i}"
-            );
-        }
-        let visible: std::collections::HashMap<u16, u16> = tree.iter().collect();
-
-        for key in 0_u16..n {
-            let i = unique_key(key);
-            let v = unique_key(key * 2);
-            assert_eq!(visible.get(&i).copied(), Some(v), "failed to get key {i}");
-        }
-
-        for key in 0..n {
-            let i = unique_key(key);
-            let v = unique_key(key * 2);
-            assert_eq!(tree.remove(&i), Some(v));
-        }
-        for key in 0..n {
-            let i = unique_key(key);
-            assert_eq!(tree.get(&i), None, "failed to get key {i}");
-        }
-    };
-
-    let tree = ConcurrentMap::default();
-
-    std::thread::scope(|s| {
-        for _ in 0..64 {
-            let barrier = std::sync::Arc::new(std::sync::Barrier::new(concurrency));
-            let mut threads = vec![];
-            for i in 0..concurrency {
-                let tree_2 = tree.clone();
-                let barrier_2 = barrier.clone();
-
-                let thread = s.spawn(move || run(tree_2, &barrier_2, u16::try_from(i).unwrap()));
-                threads.push(thread);
-            }
-            for thread in threads {
-                thread.join().unwrap();
-            }
-        }
-    });
-}
-
-#[test]
-fn big_scan() {
-    let n: u32 = 16 * 1024 * 1024;
-    let concurrency = 8;
-    let stride = n / concurrency;
-
-    let fill =
-        |tree: ConcurrentMap<u32, u32>, barrier: &std::sync::Barrier, start_fill, stop_fill| {
-            barrier.wait();
-            let insert = std::time::Instant::now();
-            for i in start_fill..stop_fill {
-                tree.insert(i, i);
-            }
-            let insert_elapsed = insert.elapsed();
-            println!(
-                "{} inserts/s, total {} in {:?}",
-                (u64::from(stride) * 1000)
-                    / u64::try_from(insert_elapsed.as_millis()).unwrap_or(u64::MAX),
-                stop_fill - start_fill,
-                insert_elapsed
-            );
-        };
-
-    let read = |tree: ConcurrentMap<u32, u32>, barrier: &std::sync::Barrier| {
-        barrier.wait();
-        let scan = std::time::Instant::now();
-        let count = tree.range(..).take(stride as _).count();
-        assert_eq!(count, stride as _);
-        let scan_elapsed = scan.elapsed();
-        let scan_micros = scan_elapsed.as_micros().max(1) as f64;
-        println!(
-            "{} scanned items/s, total {:?}",
-            ((f64::from(stride) * 1_000_000.0) / scan_micros) as u64,
-            scan_elapsed
-        );
-    };
-
-    let tree = ConcurrentMap::default();
-    let barrier = std::sync::Barrier::new(concurrency as _);
-
-    std::thread::scope(|s| {
-        let mut threads = vec![];
-        for i in 0..concurrency {
-            let tree_2 = tree.clone();
-            let barrier_2 = &barrier;
-
-            let start_fill = i * stride;
-            let stop_fill = (i + 1) * stride;
-
-            let thread = s.spawn(move || fill(tree_2, barrier_2, start_fill, stop_fill));
-            threads.push(thread);
-        }
-        for thread in threads {
-            thread.join().unwrap();
-        }
-    });
-
-    std::thread::scope(|s| {
-        let mut threads = vec![];
-        for _ in 0..concurrency {
-            let tree_2 = tree.clone();
-            let barrier_2 = &barrier;
-
-            let thread = s.spawn(move || read(tree_2, barrier_2));
-            threads.push(thread);
-        }
-        for thread in threads {
-            thread.join().unwrap();
-        }
-    });
-}
-
-#[test]
-fn bulk_load() {
-    let n: u64 = 16 * 1024 * 1024;
-
-    let concurrency = std::thread::available_parallelism()
-        .map(std::num::NonZeroUsize::get)
-        .unwrap_or(8) as u64;
-
-    let run = |tree: ConcurrentMap<u64, u64>, barrier: &std::sync::Barrier, low_bits| {
-        let shift = concurrency.next_power_of_two().trailing_zeros();
-        let unique_key = |key| (key << shift) | low_bits;
-
-        barrier.wait();
-        for key in 0..n / concurrency {
-            let i = unique_key(key);
-            tree.insert(i, i);
-        }
-    };
-
-    let tree = ConcurrentMap::default();
-
-    std::thread::scope(|s| {
-        let barrier = std::sync::Arc::new(std::sync::Barrier::new(1 + concurrency as usize));
-        let mut threads = vec![];
-        for i in 0..concurrency {
-            let tree_2 = tree.clone();
-            let barrier_2 = barrier.clone();
-
-            let thread = s.spawn(move || run(tree_2, &barrier_2, i));
-            threads.push(thread);
-        }
-        barrier.wait();
-        let insert = std::time::Instant::now();
-        for thread in threads {
-            thread.join().unwrap();
-        }
-        let insert_elapsed = insert.elapsed();
-        println!(
-            "{} bulk inserts/s, total {:?}",
-            (n * 1000) / u64::try_from(insert_elapsed.as_millis()).unwrap_or(u64::MAX),
-            insert_elapsed
-        );
-    });
 }
